+++ .gitignore
@@ -38,3 +38,5 @@ target/
 
 # Programatically generated files
 generate_pylist.py
+
+node_modules/
