+++ analyses/driver/matcher.go
@@ -58,11 +58,6 @@ type ErrorResponse struct {
 	Error string `json:"error"`
 }
 
-type StackAnalysisResponse struct {
-	HasVul bool
-	Error  error
-}
-
 // ReadManifestResponse is arg type of readManifest func
 type ReadManifestResponse struct {
 	DepsTreePath     string `json:"manifest,omitempty"`
